refactor: create folders and clean up index.ts
