+++ src/oxlint/rules.ts
@@ -4,8 +4,8 @@ import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { execSync } from "node:child_process";
 import { exit, argv } from "node:process";
-import type { OxlintRule, OxlintConfig, RuleStatus, RulesState } from "./types.js";
-import { OXLINT_VERSION } from "./index.js";
+import type { OxlintRule, OxlintConfig, RuleStatus, RulesState } from "../types.js";
+import { getState } from "../state.js";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 export function loadDescriptions() {
@@ -35,7 +35,7 @@ function getRuleStatus(ruleName: string, category: string, config: OxlintConfig)
 
 function fetchRulesFromOxlint(): any[] {
   try {
-    const raw = execSync(`npx -q --yes oxlint@${OXLINT_VERSION} --rules --format=json`, {
+    const raw = execSync(`npx -q --yes oxlint@${getState().oxlintVersion} --rules --format=json`, {
       encoding: "utf8",
       stdio: ["ignore", "pipe", "ignore"],
     });
